Rename run_tests to runner.
The name run_tests implies that it can be executed/ it can run tests.
It actually only contains the runner.
